feat: add Authentication Controller and refactor User Controller
